First Version Programmed
The arguments of the agent are given by reference, so every modification
an agent do on them are visible to the other agents. Therefore, if 
all the agents receive a Map, the evolution of the roads in the map,
detected and modified by segmentAgents, are shared between all the
agents without the need of providing it by means of a message.
The map is shared between all the agents and its dynamic is
automatically shared among them.
This simplifies a lot the problem of maintaining the same data structure
among all the agents, and the actions need to provide this feature by
sending and receiving messages.