New version running
TODO: A version in which each car has its own jgrapht. If the car is
smart new modification on weights are provided by the segment agent in
which the car is moving when the car is close to cover the segment.
TODO. when a segment changes its weight (because a change in 
its level of service was done, it sends to the other segments
this change in order to all the segments have the same 
information on weights for the traffic.